Cleanup diff
